requirements.txt
discord.py==2.4.0
requests==2.25.1
pandas==2.2.2
ta==0.11.0
python-dotenv==0.19.2
